Basic logic do decide if work apps or private apps.

+++ Windows/Install-Apps.ps1
@@ -1,6 +1,15 @@
 # Installs New applications.
 # Source: https://chrislayers.com/2021/08/01/scripting-winget/.
 
+param (
+    [Parameter(Mandatory)]
+    [bool]
+    $IsWork,
+
+    [switch]
+    $DryRun
+)
+
 $apps = @(
     # Development.
     @{name = "Microsoft.PowerShell" }
@@ -22,18 +31,25 @@ $apps = @(
     @{name = "Bitwarden.Bitwarden" }
     @{name = "voidtools.Everything" }
 
-    # Private Only.
-    # @{name = "Microsoft.VisualStudio.2022.Community" }
-    # @{name = "Foxit.FoxitReader" }
-    # @{name = "Google.Drive" }
-    # @{name = "Foundry376.Mailspring" }
-    # @{name = "BraveSoftware.BraveBrowser" }
-    # @{name = "OpenWhisperSystems.Signal" }
-    # @{name = "WhatsApp.WhatsApp" }
-    # @{name = "Twilio.Authy" }
-    # @{name = "Logitech.GHUB" }
+    # Misc.
+    @{name = "Valve.Steam" }
+    @{name = "EpicGames.EpicGamesLauncher" }
+    @{name = "Spotify.Spotify" }
+);
 
-    # Work.
+$privateOnlyApps = @(
+    @{name = "Microsoft.VisualStudio.2022.Community" }
+    @{name = "Foxit.FoxitReader" }
+    @{name = "Google.Drive" }
+    @{name = "Foundry376.Mailspring" }
+    @{name = "BraveSoftware.BraveBrowser" }
+    @{name = "OpenWhisperSystems.Signal" }
+    @{name = "WhatsApp.WhatsApp" }
+    @{name = "Twilio.Authy" }
+    @{name = "Logitech.GHUB" }
+);
+
+$workOnlyApps = @(
     # This will throw an error but it is actually a success that only requires a system restart.
     @{name = "SlackTechnologies.Slack" }
     @{name = "Microsoft.Teams" }
@@ -43,26 +59,37 @@ $apps = @(
     @{name = "mRemoteNG.mRemoteNG" }
     @{name = "Microsoft.AzureStorageExplorer" }
     @{name = "Microsoft.VisualStudio.2022.Professional" }
-
-    # Misc.
-    @{name = "Valve.Steam" }
-    @{name = "EpicGames.EpicGamesLauncher" }
-    @{name = "Spotify.Spotify" }
 );
 
-Foreach ($app in $apps) {
+if ($IsWork) {
+    $apps += $workOnlyApps
+} else {
+    $apps += $privateOnlyApps
+}
+
+Foreach ($app in $apps)
+{
+    Write-Host($app.name)
+    if ($DryRun) {
+        continue
+    }
+
     # Checks if the app is already installed.
     $listApp = winget list --exact -q $app.name
-    if (![String]::Join("", $listApp).Contains($app.name)) {
+    if (![String]::Join("", $listApp).Contains($app.name))
+    {
         Write-host "Installing:" $app.name
-        if ($null -ne $app.source) {
+        if ($null -ne $app.source)
+        {
             winget install --exact --silent $app.name --source $app.source
         }
-        else {
+        else
+        {
             winget install --exact --silent $app.name
         }
     }
-    else {
+    else
+    {
         Write-host "Skipping Install of " $app.name
     }
 }
